V0.07.alpha (#312)
* Fixed missing dsl_tiler_tiles_get() and dsl_tiler_tiles_set() (#282)

* WIP Implement Record Tap (#284)

* Implement Record Tap (#287)

* WIP Implement Record Tap

* Implement Record Tap

* Fixed Fake Sink LinkAll/UnlinkAll (#288)

* WIP New Area Types (#292)

* resolve throughput issues (#297)

* Implement Meter Sink to measure Pipeline throughput (#299)

* WIP Implement Meter Sink

* Implement Meter Sink

* Meter sink (#300)

* WIP Implement Meter Sink

* Implement Meter Sink

* Complete Meter Sink implementation

* Implement Tiler Show Source Services (#302)

* Fix client_data always null on callback (#306)

* Added latency paramter for RTSP source (#304)

* Refactor ODE Handler component and Meter Sink to new Pad-Probe-Handler (#309)

* WIP Refactor ODE Handler and Meter

* WIP Refactor ODE Handler and Meter Sink

* Refactor ODE Handler and Sink Metter to New PPH type

* Update Tiler Show Source (#311)

+++ test/api/DslOdeTriggerApiTest.cpp
@@ -158,14 +158,14 @@ SCENARIO( "An ODE Trigger's minimum dimensions can be set/get", "[ode-trigger-ap
 
         REQUIRE( dsl_ode_trigger_occurrence_new(odeTriggerName.c_str(), class_id, limit) == DSL_RESULT_SUCCESS );
 
-        uint min_width(1), min_height(1);
+        float min_width(1), min_height(1);
         REQUIRE( dsl_ode_trigger_dimensions_min_get(odeTriggerName.c_str(), &min_width, &min_height) == DSL_RESULT_SUCCESS );
         REQUIRE( min_width == 0 );
         REQUIRE( min_height == 0 );
 
         WHEN( "When the Trigger's min dimensions are updated" )         
         {
-            uint new_min_width(300), new_min_height(200);
+            float new_min_width(300), new_min_height(200);
             REQUIRE( dsl_ode_trigger_dimensions_min_set(odeTriggerName.c_str(), new_min_width, new_min_height) == DSL_RESULT_SUCCESS );
             
             THEN( "The correct value is returned on get" ) 
@@ -190,7 +190,7 @@ SCENARIO( "An ODE Trigger's maximum dimensions can be set/get", "[ode-trigger-ap
 
         REQUIRE( dsl_ode_trigger_occurrence_new(odeTriggerName.c_str(), class_id, limit) == DSL_RESULT_SUCCESS );
 
-        uint max_width(1), max_height(1);
+        float max_width(1), max_height(1);
         REQUIRE( dsl_ode_trigger_dimensions_max_get(odeTriggerName.c_str(), &max_width, &max_height) == DSL_RESULT_SUCCESS );
         REQUIRE( max_width == 0 );
         REQUIRE( max_height == 0 );
